pyproject.toml
[project]
name = "goit-pythonweb-hw-06"
version = "0.1.0"
description = ""
authors = [
    {name = "VOSolyanik",email = "v.o.solyanik@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "alembic (>=1.16.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "faker (>=37.4.0,<38.0.0)",
    "greenlet (>=3.2.3,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

